+++ gdeep/trainer/regularizer.py
@@ -1,4 +1,4 @@
-from gdeep.trainer import Trainer
+#from gdeep.trainer import Trainer
 import os
 import copy
 import time
